* Use int instead of short for scores (it can be pretty big cause we use it to store the time in ms) * Fix small bug : only update score when beating personal best

+++ source/gameshared/q_comref.h
@@ -430,7 +430,8 @@ typedef struct
 	qbyte weaponState;
 
 	int inventory[MAX_ITEMS];
-	short stats[PS_MAX_STATS];	// fast status bar updates
+	// racesow : use int for scores cause we use it for time
+	int stats[PS_MAX_STATS];	// fast status bar updates
 	qbyte plrkeys;				// infos on the pressed keys of chased player (self if not chasing)
 } player_state_t;
